+++ WorldCities/WorldCities/src/app/app.module.ts
@@ -10,7 +10,8 @@ import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
 import { NavMenuComponent } from './nav-menu/nav-menu.component';
 import { CitiesComponent } from './cities/cities.component';
-import { CountriesComponent } from './countries/countries.component'
+import { CountriesComponent } from './countries/countries.component';
+import { CityEditComponent } from './cities/city-edit.component';
 
 @NgModule({
   declarations: [
@@ -18,7 +19,8 @@ import { CountriesComponent } from './countries/countries.component'
     HomeComponent,
     NavMenuComponent,
     CitiesComponent,
-    CountriesComponent
+    CountriesComponent,
+    CityEditComponent
   ],
   imports: [
     BrowserModule,